requirements.txt
streamlit
pandas
plotly
google-genai
tabulate
openpyxl
requests